+++ src/Business/Grand.Business.Catalog/Services/Tax/TaxService.cs
@@ -375,7 +375,7 @@ public class TaxService : ITaxService
         Customer customer,
         double unitPrice,
         double unitPriceWithoutDisc,
-        int quantity,
+        double quantity,
         double subTotal,
         double discountAmount,
         bool priceIncludesTax
