+++ src/Web/Grand.Web.Admin/Interfaces/IProductViewModelService.cs
@@ -11,7 +11,7 @@ public interface IProductViewModelService
     Task PrepareProductReviewModel(ProductReviewModel model, ProductReview productReview, bool excludeProperties,
         bool formatReviewText);
 
-    Task OutOfStockNotifications(Product product, ProductModel model, int prevStockQuantity,
+    Task OutOfStockNotifications(Product product, ProductModel model, double prevStockQuantity,
         List<ProductWarehouseInventory> prevMultiWarehouseStock);
 
     Task OutOfStockNotifications(Product product, ProductAttributeCombination combination,
