+++ src/Web/Grand.Web.Admin/Interfaces/IShipmentViewModelService.cs
@@ -10,8 +10,8 @@ public interface IShipmentViewModelService
     Task<ShipmentModel> PrepareShipmentModel(Shipment shipment, bool prepareProducts,
         bool prepareShipmentEvent = false);
 
-    Task<int> GetStockQty(Product product, string warehouseId);
-    Task<int> GetReservedQty(Product product, string warehouseId);
+    Task<double> GetStockQty(Product product, string warehouseId);
+    Task<double> GetReservedQty(Product product, string warehouseId);
     Task<IList<ShipmentModel.ShipmentNote>> PrepareShipmentNotes(Shipment shipment);
     Task InsertShipmentNote(Shipment shipment, string downloadId, bool displayToCustomer, string message);
     Task DeleteShipmentNote(Shipment shipment, string id);
